+++ src/Collector/Schema/SchemaUsage.php
@@ -10,6 +10,13 @@ use PhpParser\Node\Expr\New_;
 use PhpParser\Node\Name;
 use PHPStan\Analyser\Scope;
 use PHPStan\Collectors\Collector;
+use function count;
+use function get_class;
+use function is_numeric;
+use function is_string;
+use function json_encode;
+use function property_exists;
+use function strpos;
 
 /**
  * @implements Collector<New_, array{string, string, int}>