+++ src/Rule/Schema/UnusedProperties.php
@@ -14,6 +14,15 @@ use PHPStan\Analyser\Scope;
 use PHPStan\Node\CollectedDataNode;
 use PHPStan\Rules\Rule;
 use PHPStan\Rules\RuleErrorBuilder;
+use function array_keys;
+use function array_unique;
+use function count;
+use function implode;
+use function is_array;
+use function json_decode;
+use function sprintf;
+use function strpos;
+use function trim;
 
 /**
  * @implements Rule<CollectedDataNode>