+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased][unreleased]
+
+## [0.8.0] - 2026-01-26
 ### Added
 - Support for Neon DI attribute Inject
 - Set PHP 8.1 as minimum version
@@ -83,7 +85,8 @@
 - Rule for checking trait context 
 - Phpstorm meta dynamic return types
 
-[unreleased]: https://github.com/efabrica-team/phpstan-rules/compare/0.7.6...HEAD
+[unreleased]: https://github.com/efabrica-team/phpstan-rules/compare/0.8.0...HEAD
+[0.8.0]: https://github.com/efabrica-team/phpstan-rules/compare/0.7.6...0.8.0
 [0.7.6]: https://github.com/efabrica-team/phpstan-rules/compare/0.7.5...0.7.6
 [0.7.5]: https://github.com/efabrica-team/phpstan-rules/compare/0.7.4...0.7.5
 [0.7.4]: https://github.com/efabrica-team/phpstan-rules/compare/0.7.3...0.7.4